Removed base class usage, added ability to suppress missing plugin warning in log
